+++ frontend/src/components/list/hooks/useListUpdate.ts
@@ -2,7 +2,7 @@ import { listState } from "components/list/state/listAtoms";
 import { List } from "gql/codegen-output";
 import { useMutateDeleteList, useMutateUpdateList } from "gql/hooks/list-mutate";
 import { DeleteTermsVariables, useMutateDeleteTerms } from "gql/hooks/term-query";
-import { useRouteProps } from "hooks/routerHooks";
+import useRouteProps from "hooks/useRouteProps";
 import { useCallback, useEffect } from "react";
 import { useSetRecoilState } from "recoil";
