+++ frontend/src/components/lists/Lists.tsx
@@ -1,4 +1,4 @@
-import { useRouteProps } from "hooks/routerHooks";
+import useRouteProps from "hooks/useRouteProps";
 import { useMemo } from "react";
 import { Link } from "react-router-dom";
 import useLists from "./hooks/useLists";
