+++ frontend/src/components/user/User.tsx
@@ -1,22 +1,22 @@
 import UserSessions from "components/sessions/UserSessions";
-import { useRouteProps } from "../../hooks/routerHooks";
+import useRouteProps from "../../hooks/useRouteProps";
 import cs from "./User.module.scss";
 
 const User = () => {
-    const { params } = useRouteProps();
-    const username = params.username;
+	const { params } = useRouteProps();
+	const username = params.username;
 
-    return (
-        <div className="PageWrapper">
-            <div className="User">
-                <div className={cs.Title}>/u/{username}'s profile</div>
-            </div>
+	return (
+		<div className="PageWrapper">
+			<div className="User">
+				<div className={cs.Title}>/u/{username}'s profile</div>
+			</div>
 
-            <section>
-                <UserSessions />
-            </section>
-        </div>
-    );
+			<section>
+				<UserSessions />
+			</section>
+		</div>
+	);
 };
 
 export default User;
